requirements.txt
numpy
pillow
matplotlib
tqdm
streamlit
spandrel
spandrel_extra_arches
streamlit-image-comparison